+++ data/cube/aggregate/AverageStrictAggregator.js
@@ -6,10 +6,12 @@
  */
 
 import {Aggregator} from '@xh/hoist/data/cube/aggregate/Aggregator';
+import {gatherLeaves} from '@xh/hoist/utils/js';
 
 export class AverageStrictAggregator extends Aggregator {
 
     aggregate(rows, fieldName) {
+        rows = gatherLeaves(rows);
         let total = null;
 
         for (const row of rows) {