+++ CHANGELOG.md
@@ -75,6 +75,8 @@
 ### 🐞 Bug Fixes
 
 * Fixed a regression with `StoreFilterField` introduced in v33.0.1.
+* Average aggregators now correctly takes an average of all leaves vs. an average of averages
+  in the case of nodes more than one level deep.
 
 [Commit Log](https://github.com/xh/hoist-react/compare/v33.0.2...33.1.0)